app/requirements.txt
absl-py==0.7.1
altair==4.1.0
app==0.0.1
appdirs==1.4.4
appnope==0.1.0
astor==0.8.1
astropy==3.1.2
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
boto3==1.14.4
botocore==1.17.4
branca==0.4.0
Brotli==1.0.7
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
classy==2.7.2
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cmdstanpy==0.4.0
colour==0.1.5
contextlib2==0.6.0.post1
convertdate==2.2.1
cycler==0.10.0
Cython==0.29.10
dash==0.21.1
dash-core-components==0.23.0
dash-html-components==0.11.0
dash-renderer==0.13.0
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
distlib==0.3.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
ephem==3.7.7.1
fake-useragent==0.1.11
filelock==3.0.12
Fiona==1.8.13.post1
Flask==1.1.2
Flask-Compress==1.5.0
folium==0.10.1
gast==0.2.2
geopandas==0.7.0
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.19.0
h5py==2.9.0
healpy==1.12.9
holidays==0.10.2
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
jellyfish==0.6.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.13.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.0.0
jupyter-core==4.6.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.9
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
llvmlite==0.32.1
LunarCalendar==0.0.9
lxml==4.5.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mido==1.2.9
mistune==0.8.4
mock==2.0.0
munch==2.5.0
music21==5.7.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numba==0.49.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
patsy==0.5.1
pbr==5.1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.8.1
plotly-geo==1.0.0
pmdarima==1.6.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.4.0b2
pygame==1.9.6
Pygments==2.6.1
PyMeeus==0.3.7
pyparsing==2.4.7
pyproj==2.6.0
pyrsistent==0.16.0
pyshp==2.1.0
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.1
pyzmq==19.0.1
qtconsole==4.4.3
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.0
selenium==3.141.0
Send2Trash==1.5.0
setuptools-git==1.2
Shapely==1.7.0
six==1.12.0
sklearn==0.0
soupsieve==2.0.1
statsmodels==0.11.1
streamlit==0.61.0
tb-nightly==1.14.0a20190301
tbb==2019.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==1.14.0.dev2019030115
tf-slim==1.1.0
threadpoolctl==2.1.0
tifffile==2020.6.3
toml==0.10.1
toolz==0.10.0
tornado==5.1.1
traitlets==4.3.3
tzlocal==2.1
umap-learn==0.4.4
urllib3==1.25.9
us==2.0.2
validators==0.15.0
virtualenv==20.0.21
watchdog==0.10.2
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==0.15.1
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
yellowbrick==1.1
zipp==3.1.0
